Rename everything from ot3_state_manager to just state_manager
Fix target names


Make it so setup does not alter the lock file


test


Not much
