+++ CMakeLists.txt
@@ -71,12 +71,12 @@ list(REMOVE_DUPLICATES LINT_TARGETS)
 
 message(CHECK_START "Checking if cross-compiling.")
 if(NOT ${CMAKE_CROSSCOMPILING})
-	message(CHECK_PASS "Not cross-compiling. Adding ot3_state_manager")
+	message(CHECK_PASS "Not cross-compiling. Adding state_manager")
 	find_package(Poetry 1.1.15)
-	add_subdirectory(ot3_state_manager)
+	add_subdirectory(state_manager)
 	message("Poetry executable is located at: ${Poetry_EXECUTABLE}")
 else ()
-	message(CHECK_FAIL "You are cross-compiling. Not adding ot3_state_manager")
+	message(CHECK_FAIL "You are cross-compiling. Not adding state_manager")
 endif ()
 
 find_package(Clang)
