--- ot3_state_manager/ot3_state_manager/__init__.py
@@ -1,1 +1,0 @@
-"""ot3_state_manager package."""
