--- ot3_state_manager/tests/__init__.py
@@ -1,1 +1,0 @@
-"""ot3_state_manager test packages."""
