+++ state_manager/state_manager/__init__.py
@@ -1,0 +1,1 @@
+"""state_manager package."""
