+++ state_manager/tests/__init__.py
@@ -1,0 +1,1 @@
+"""state_manager test packages."""
