+++ state_manager/tests/conftest.py
@@ -2,8 +2,8 @@
 
 import pytest
 
-from ot3_state_manager.ot3_state import OT3State
-from ot3_state_manager.pipette_model import PipetteModel
+from state_manager.ot3_state import OT3State
+from state_manager.pipette_model import PipetteModel
 
 
 @pytest.fixture
